refactor. 테스트용 로또 생성기 외부 클래스로 분할
