+++ src/test/java/AutoLottoTicketTest.java
@@ -1,12 +1,7 @@
 import model.AutoCount;
 import model.AutoLottoTicket;
-import model.Lotto;
-import model.LottoNumber;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import util.LottoGenerator;
-
-import java.util.List;
+import util.TestLottoGenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -20,25 +15,5 @@ public class AutoLottoTicketTest {
         AutoLottoTicket ticket = new AutoLottoTicket(generator, count);
 
         assertThat(ticket.getLottos()).hasSize(3);
-    }
-
-    @Nested
-    public class TestLottoGenerator implements LottoGenerator {
-
-        public TestLottoGenerator() { }
-
-        @Override
-        public Lotto generate() {
-            return new Lotto(
-                List.of(
-                    new LottoNumber(1),
-                    new LottoNumber(2),
-                    new LottoNumber(3),
-                    new LottoNumber(4),
-                    new LottoNumber(5),
-                    new LottoNumber(6)
-                )
-            );
-        }
     }
 }
